running the app now displays org homepage. have NOT tested DB stuff since im now using FirebaseConnector class (should still be fine)
