+++ app/src/main/java/com/example/plannet/ui/dashboard/DashboardFragment.java
@@ -1,4 +1,4 @@
-package com.example.plannet.ui.Fragments.events;
+package com.example.plannet.ui.dashboard;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.plannet.databinding.FragmentEventsBinding;
+import com.example.plannet.databinding.FragmentDashboardBinding;
 
-public class EventsFragment extends Fragment {
+public class DashboardFragment extends Fragment {
 
-    private FragmentEventsBinding binding;
+    private FragmentDashboardBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        EventsViewModel eventsViewModel =
-                new ViewModelProvider(this).get(EventsViewModel.class);
+        DashboardViewModel dashboardViewModel =
+                new ViewModelProvider(this).get(DashboardViewModel.class);
 
-        binding = FragmentEventsBinding.inflate(inflater, container, false);
+        binding = FragmentDashboardBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textEvents;
-        eventsViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textDashboard;
+        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
