+++ app/src/main/java/com/example/plannet/ui/events/EventsViewModel.java
@@ -1,16 +1,19 @@
-package com.example.plannet.ui.Fragments.home;
+package com.example.plannet.ui.events;
+
+import android.util.Log;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class EventsViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public EventsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("Events go here!!");
+        Log.d("EventsViewModel", "Text value: " + mText.getValue());
     }
 
     public LiveData<String> getText() {
