+++ app/src/main/java/com/example/plannet/ui/notifications/NotificationsFragment.java
@@ -1,4 +1,4 @@
-package com.example.plannet.ui.Fragments.notifications;
+package com.example.plannet.ui.notifications;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
