+++ app/src/main/res/layout/fragment_home.xml
@@ -4,21 +4,22 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Fragments.home.HomeFragment"
+    tools:context=".ui.home.HomeFragment"
     android:background="@color/backgroundColor">
 
     <TextView
         android:id="@+id/title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="44dp"
-        android:layout_marginTop="16dp"
         android:fontFamily="@font/poppins"
         android:text="Facility: FACILITYNAME"
         android:textColor="@color/white"
         android:textSize="30sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.023" />
 
     <ImageView
         android:id="@+id/imageView7"
@@ -30,45 +31,50 @@
         app:layout_constraintTop_toBottomOf="@id/title"/>
 
     <Button
-        android:id="@+id/button_new_event"
+        android:id="@+id/buttonNewEvent"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="72dp"
         android:layout_marginTop="220dp"
         android:backgroundTint="@color/buttonPurple"
         android:fontFamily="@font/poppins"
         android:text="Create a new event"
         android:textAllCaps="false"
         android:textSize="20sp"
+        app:layout_constraintBottom_toTopOf="@+id/buttonQrCodes"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.496"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.791"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <Button
-        android:id="@+id/button_qr_codes"
+        android:id="@+id/buttonQrCodes"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginBottom="380dp"
         android:backgroundTint="@color/buttonPurple"
         android:fontFamily="@font/poppins"
         android:text="Hashed QR Codes"
         android:textAllCaps="false"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.473" />
 
     <Button
-        android:id="@+id/button_draw"
+        android:id="@+id/buttonDraw"
         android:layout_width="279dp"
         android:layout_height="59dp"
-        android:layout_marginStart="64dp"
-        android:layout_marginBottom="296dp"
         android:backgroundTint="@color/buttonPurple"
         android:fontFamily="@font/poppins"
         android:text="Draw Attendees for Events"
         android:textAllCaps="false"
         android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.572" />
 </androidx.constraintlayout.widget.ConstraintLayout>
