+++ app/src/main/res/layout/fragment_organizer_create_event.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Fragments.home.OrganizerCreateEventFragment">
+    tools:context=".OrganizerCreateEventFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
